+++ src/array/codec/array_to_bytes/codec_chain.rs
@@ -9,7 +9,7 @@ use crate::{
             CodecTraits,
         },
         concurrency::RecommendedConcurrency,
-        ArrayView, BytesRepresentation, ChunkRepresentation, ChunkShape,
+        ArrayMetadataOptions, ArrayView, BytesRepresentation, ChunkRepresentation, ChunkShape,
     },
     metadata::Metadata,
     plugin::PluginCreateError,
@@ -130,23 +130,29 @@ impl CodecChain {
 
     /// Create codec chain metadata.
     #[must_use]
-    pub fn create_metadatas(&self) -> Vec<Metadata> {
+    pub fn create_metadatas_opt(&self, options: &ArrayMetadataOptions) -> Vec<Metadata> {
         let mut metadatas =
             Vec::with_capacity(self.array_to_array.len() + 1 + self.bytes_to_bytes.len());
         for codec in &self.array_to_array {
-            if let Some(metadata) = codec.create_metadata() {
+            if let Some(metadata) = codec.create_metadata_opt(options) {
                 metadatas.push(metadata);
             }
         }
-        if let Some(metadata) = self.array_to_bytes.create_metadata() {
+        if let Some(metadata) = self.array_to_bytes.create_metadata_opt(options) {
             metadatas.push(metadata);
         }
         for codec in &self.bytes_to_bytes {
-            if let Some(metadata) = codec.create_metadata() {
+            if let Some(metadata) = codec.create_metadata_opt(options) {
                 metadatas.push(metadata);
             }
         }
         metadatas
+    }
+
+    /// Create codec chain metadata with default options.
+    #[must_use]
+    pub fn create_metadatas(&self) -> Vec<Metadata> {
+        self.create_metadatas_opt(&ArrayMetadataOptions::default())
     }
 
     /// Get the array to array codecs
@@ -201,8 +207,8 @@ impl CodecChain {
 impl CodecTraits for CodecChain {
     /// Returns [`None`] since a codec chain does not have standard codec metadata.
     ///
-    /// Note that usage of the codec chain is explicit in [`Array`](crate::array::Array) and [`CodecChain::create_metadatas()`] will call [`CodecTraits::create_metadata()`] from for each codec.
-    fn create_metadata(&self) -> Option<Metadata> {
+    /// Note that usage of the codec chain is explicit in [`Array`](crate::array::Array) and [`CodecChain::create_metadatas_opt()`] will call [`CodecTraits::create_metadata_opt()`] from for each codec.
+    fn create_metadata_opt(&self, _options: &ArrayMetadataOptions) -> Option<Metadata> {
         None
     }