+++ src/array/codec/array_to_bytes/zfp/zfp_codec.rs
@@ -13,7 +13,7 @@ use crate::{
             BytesPartialDecoderTraits, CodecError, CodecOptions, CodecTraits,
             RecommendedConcurrency,
         },
-        BytesRepresentation, ChunkRepresentation, DataType,
+        ArrayMetadataOptions, BytesRepresentation, ChunkRepresentation, DataType,
     },
     metadata::Metadata,
 };
@@ -97,7 +97,7 @@ impl ZfpCodec {
 }
 
 impl CodecTraits for ZfpCodec {
-    fn create_metadata(&self) -> Option<Metadata> {
+    fn create_metadata_opt(&self, _options: &ArrayMetadataOptions) -> Option<Metadata> {
         let configuration = match self.mode {
             ZfpMode::Expert(expert) => ZfpCodecConfigurationV1::Expert(expert),
             ZfpMode::FixedRate(rate) => {