+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Add "experimental codec store metadata if encode only" option to global config
+ - Add `ArrayMetadataOptions`
+
 ### Changed
  - **Breaking**: Bump MSRV to 1.73 (5 October, 2023)
  - Bump `pco` (pcodec) to 0.2.1
+ - **Breaking**: Add `CodecTraits::create_metadata_opt()`
 
 ## [0.12.5] - 2024-03-17